+++ R/sheets_delete.R
@@ -2,7 +2,7 @@
 #'
 #' Deletes a range of cells and shifts other cells into the deleted area. There
 #' are several related tasks that are done by other functions:
-#'   * To clear cells of their value and/or format, use [sheets_clear()].
+#'   * To clear cells of their value and/or format, use [range_clear()].
 #'   * To delete an entire (work)sheet, use [sheets_sheet_delete()].
 #'   * To change the dimensions of a (work)sheet, use [sheets_sheet_resize()].
 #'