+++ NEWS.md
@@ -11,8 +11,8 @@ These functions are ready for use but are still considered experimental and may 
     existing (spread)Sheet.
   * `sheets_append()` adds rows to the data in an existing sheet.
   * `sheets_edit()` writes to a range.
-  * `sheets_flood()` "floods" all cells in a range with the same content.
-    `sheets_clear()` is a wrapper around `sheets_flood()` for the special case
+  * `range_flood()` "floods" all cells in a range with the same content.
+    `range_clear()` is a wrapper around `range_flood()` for the special case
      of clearing cell values.
   * `sheets_delete()` deletes a range of cells.